+++ time.hpp
@@ -4,12 +4,12 @@
 struct Time
 {
 public:
-  static Time fromSeconds(double seconds);
+  static Time fromSeconds(float seconds);
   static Time fromMilliseconds(uint32_t milliseconds);
   static Time now();
   static void reset();
 
-  double getSeconds() const;
+  float getSeconds() const;
   uint32_t getMilliseconds() const;
 
 private:
@@ -17,7 +17,7 @@ private:
   static uint32_t baseline;
 };
 
-double operator/(const Time &time1, const Time &time2);
+float operator/(const Time &time1, const Time &time2);
 Time operator+(const Time &time1, const Time &time2);
 Time operator-(const Time &time1, const Time &time2);
 bool operator<(const Time &time1, const Time &time2);
